Add files via upload
Hi everyone, I saw that there are puzzles in EDOPro from the old DS game Yu-Gi-Oh! NightmareTroubadour. So I rebuild all puzzles from Yu-Gi-Oh! GX Spirit Caller (2007), 
thats the sequel to 2005's Nightmare Troubadour.
They all work perfectly (especially because of Naims, Edos and PyrQs help^^). I've added complexity-ratings and solutions for each puzzle. 
45 puzzles in total. Divided into beginner, intermediate and advanced just like in the original DS game.

I thought it would be great to make it accessible to everyone. Could it be implemented ingame? (Canon collection)
It would be a shame if people didn't have access to them, because they are high-quality puzzles. Is there a possibility?
Suggestions for improvement are welcome. 
Thanks in advance.